+++ src/compiler/copying.c
@@ -762,7 +762,6 @@ RETRY:
 			MACRO_COPY_ASTID(ast->return_stmt.cleanup_fail);
 			break;
 		case AST_SWITCH_STMT:
-		case AST_IF_CATCH_SWITCH_STMT:
 			copy_reg_ref(c, source, ast);
 			SCOPE_FIXUP_START
 				copy_flow(c, ast);